fix: split mock data
