+++ apps/web/src/entities/exercisePreset/api/index.ts
@@ -15,14 +15,3 @@ export {
   useGetExercisePresetWithListByOffset,
   useGetExercisePresetWithListById,
 }
-
-export const ExercisePresetMockData: { [key: number]: ExercisePresetStoreType } = Array(20).fill(0).reduce((acc, _cur, i) => {
-  const id = i + 1
-  acc[id] = {
-    id: id,
-    name: `TestPreset - ${id}`,
-    deps: 0,
-    __typename: 'ExercisePreset'
-  } as ExercisePresetStoreType
-  return acc
-}, {})
