+++ apps/web/src/entities/schedule/api/index.ts
@@ -1,5 +1,3 @@
-import { ScheduleType } from '@entities/schedule/model/ScheduleType'
-import { DateService } from '@widgets/calender'
 import useCloneSchedule from './useCloneSchedule'
 import useCloneScheduleFromPreset from './useCloneScheduleFromPreset'
 import useCreateSchedule from './useCreateSchedule'
@@ -19,22 +17,3 @@ export {
   useCloneSchedule,
   useCloneScheduleFromPreset
 }
-const { month, year } = DateService.takeTodayDateValue()
-const daysByMonth = DateService.getDaysInMonth(year, month)
-
-const typeTemp = [ScheduleType.BREAK, ScheduleType.FINISH, ScheduleType.PAUSED, ScheduleType.SCHEDULED, ScheduleType.STARTED]
-export const ScheduleMockData: { [key: number]: ScheduleStoreType } = Array(daysByMonth).fill(0).reduce((acc, _cur, i) => {
-  const id = i + 1
-  acc[id] = {
-    date: id,
-    month,
-    year,
-    id: id,
-    type: typeTemp[i % 5],
-    beforeTime: 0,
-    breakTime: 0,
-    start: 0,
-    workoutTimes: 0
-  } as ScheduleStoreType
-  return acc
-}, {})
