+++ apps/web/src/entities/schedule/api/useCloneScheduleFromPreset.mock.ts
@@ -1,8 +1,8 @@
 import type { MockedResponse } from '@apollo/client/testing';
 import type { ExercisePreset } from 'fitness-struct'
-import { ExercisePresetMockData } from '@entities/exercisePreset/api';
-import { ScheduleMockData } from '@entities/schedule/api';
+import { ExercisePresetMockData } from '@entities/exercisePreset/api/exercisePreset.mockData';
 import CloneScheduleFromPreset from '@entities/schedule/api/graphql/mutation/CloneScheduleFromPresetGql';
+import { ScheduleMockData } from '@entities/schedule/api/schedule.mockData';
 
 const CloneScheduleFromPresetMock: MockedResponse<CloneScheduleFromPresetResponse, CloneScheduleFromPresetVariable> = {
   request: {
