+++ apps/web/src/entities/schedule/api/useCreateSchedule.mock.ts
@@ -1,6 +1,6 @@
 import type { MockedResponse } from '@apollo/client/testing';
-import { ScheduleMockData } from '@entities/schedule/api';
 import CreateSchedule from '@entities/schedule/api/graphql/mutation/CreateScheduleGql';
+import { ScheduleMockData } from '@entities/schedule/api/schedule.mockData';
 
 const CreateScheduleMock: MockedResponse<CreateScheduleResponse, CreateScheduleVariable> = {
   request: {
