+++ apps/web/src/entities/schedule/api/useGetScheduleStatusByMonth.mock.ts
@@ -1,7 +1,7 @@
 import type { MockedResponse } from '@apollo/client/testing';
 import type { Schedule } from 'fitness-struct';
-import { ScheduleMockData } from '@entities/schedule/api';
 import GetScheduleStatusByMonthGql from '@entities/schedule/api/graphql/query/GetScheduleStatusByMonthGql';
+import { ScheduleMockData } from '@entities/schedule/api/schedule.mockData';
 
 const GetScheduleStateByDateMock: MockedResponse<
   GetScheduleStatusByMonthResponse,
