+++ apps/web/src/entities/set/api/useDeleteSet.mock.ts
@@ -1,6 +1,6 @@
 import type { MockedResponse } from '@apollo/client/testing';
-import { SetMockData } from '@entities/set/api';
 import DeleteSetGql from '@entities/set/api/graphql/mutation/DeleteSetGql';
+import { SetMockData } from '@entities/set/api/set.mockData';
 
 const DeleteSetMock: MockedResponse<DeleteSetResponse, DeleteSetVariable> = {
   request: {
